restructure FWCore->k4FWCore
