+++ k4FWCore/src/KeepDropSwitch.cpp
@@ -1,4 +1,4 @@
-#include "FWCore/KeepDropSwitch.h"
+#include "k4FWCore/KeepDropSwitch.h"
 
 #include <iostream>
 #include <sstream>
