+++ k4FWCore/src/PodioDataSvc.cpp
@@ -1,9 +1,9 @@
-#include "FWCore/PodioDataSvc.h"
+#include "k4FWCore/PodioDataSvc.h"
 #include "GaudiKernel/IConversionSvc.h"
 #include "GaudiKernel/IEventProcessor.h"
 #include "GaudiKernel/ISvcLocator.h"
 
-#include "FWCore/DataWrapper.h"
+#include "k4FWCore/DataWrapper.h"
 
 #include "TTree.h"
